Add separate AO and normal test blocks
