+++ src/main/resources/assets/gtnhlib/blockstates/model_test.json
@@ -1,8 +1,5 @@
 {
   "variants": {
-    "facing=north": { "model": "gtnhlib:blocks/old_stairs", "y": 270, "uvlock": true },
-    "facing=south": { "model": "gtnhlib:blocks/old_stairs", "y": 90, "uvlock": true },
-    "facing=west": { "model": "gtnhlib:blocks/old_stairs", "y": 0, "uvlock": true },
-    "facing=east": { "model": "gtnhlib:blocks/old_stairs", "y": 180, "uvlock": true }
+    "": { "model": "gtnhlib:blocks/model_test" }
   }
 }
